#  Make MetricFamilyToText allocation-free