@mxinden as promised.
@grobie we should run a kube-state-metrics instance with this vendored in.
Fixes #146.

This makes use of interface upgrades of the `io.Writer`. If the
necessary upgrades are not available, a `bytes.Buffer` is used in
between.

Number formatting is performed with the `strconv.Append...` functions,
using an intermediate byte slice as buffer. (Perhaps surpsingly, a
plain `fmt.Fprint` with an int or float as single argument causes
quite a bit of allocations, despit the usage of `sync.Pool` by the
`fmt` package.)

Both the `bytes.Buffer` and the intermediate `[]byte` buffer are
managed by a `sync.Pool` each.

Benchmark comparison:

BEFORE
```
$ go test -bench=Create -run=NONE -benchmem
goos: linux
goarch: amd64
pkg: github.com/prometheus/common/expfmt
BenchmarkCreate-4                  50000             23854 ns/op            4433 B/op        285 allocs/op
BenchmarkCreateBuildInfo-4       1000000              1592 ns/op             216 B/op         19 allocs/op
```

AFTER
```
$ go test -bench=Create -run=NONE -benchmem
goos: linux
goarch: amd64
pkg: github.com/prometheus/common/expfmt
BenchmarkCreate-4                 100000             12644 ns/op               0 B/op          0 allocs/op
BenchmarkCreateBuildInfo-4       2000000               892 ns/op               0 B/op          0 allocs/op
```

Bottom line: About twice as fast, and no allocations at all in the long run.