@fabxc 2018-10-16
@beorn7 have you seen https://godoc.org/golang.org/x/tools/cmd/benchcmp?
It's pretty nice for comparing benchmark outputs.