@beorn7 2018-10-16
I know the tool, but I found it a bit over-tech'd for the six deltas involved here, especially given that four of them went down to zero...